requirements.txt

# general modules
django==4.2.20
djangorestframework==3.15.2
python-decouple==3.8
Pillow==10.4.0

# third party modules
markdown==3.4.3
django-filter==23.2
drf-yasg[validation]
djangorestframework_simplejwt==5.3.1

# deployment modules


# email third party modules
django-mail-templated==2.6.5
black==25.11.0